+++ teta_widgets/lib/src/elements/code/advanced/gridview_builder.dart
@@ -37,7 +37,7 @@ class GridViewBuilderCodeTemplate {
     final valueCrossAxisCount =
         (body.attributes[DBKeys.crossAxisCount] as FTextTypeInput).toCode(
       loop,
-      resultType: ResultTypeEnum.double,
+      resultType: ResultTypeEnum.int,
       defaultValue: '1',
     );
     final crossAxisCount = int.tryParse(valueCrossAxisCount) != null &&
